ai/FastAPI/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.27.0

# HTML/Data extraction
beautifulsoup4==4.12.2
mlscraper==0.1.0
extruct==0.14.0

# AWS
boto3==1.34.41
pyngrok==7.0.0

# Database
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2

# Image processing
opencv-python==4.9.0.80
pillow==10.2.0
rembg==2.0.50
ultralytics==8.1.13  # For YOLOv8
numpy==1.26.3

# AI models
open-clip-torch==2.20.0
torch==2.1.2
torchvision==0.16.2
transformers==4.37.2
openai==1.12.0